Fix some issues caused by i18n addition
