+++ src/renderer/modals/login/loginModal.tsx
@@ -13,8 +13,6 @@ import {AuthenticationResponseDetails, AuthInfo} from 'electron/renderer';
 import urlUtils from 'common/utils/url';
 import {MODAL_INFO} from 'common/communication';
 
-import IntlProvider from 'renderer/intl_provider';
-
 type Props = {
     onCancel: (request: AuthenticationResponseDetails) => void;
     onLogin: (request: AuthenticationResponseDetails, username: string, password: string) => void;
@@ -116,98 +114,96 @@ class LoginModal extends React.PureComponent<Props, State> {
         const {intl} = this.props;
 
         return (
-            <IntlProvider>
-                <Modal
-                    show={Boolean(this.state.request && this.state.authInfo)}
-                >
-                    <Modal.Header>
-                        <Modal.Title>
-                            <FormattedMessage
-                                id='renderer.modals.login.loginModal.title'
-                                defaultMessage='Authentication Required'
-                            />
-                        </Modal.Title>
-                    </Modal.Header>
-                    <Modal.Body>
-                        <p>
-                            {this.renderLoginModalMessage()}
-                        </p>
-                        <Form
-                            onSubmit={this.handleSubmit}
-                        >
-                            <FormGroup>
-                                <Col
-                                    as={FormLabel}
-                                    sm={2}
-                                >
-                                    <FormattedMessage
-                                        id='renderer.modals.login.loginModal.username'
-                                        defaultMessage='User Name'
-                                    />
-                                </Col>
-                                <Col sm={10}>
-                                    <FormControl
-                                        type='text'
-                                        placeholder={intl.formatMessage({id: 'renderer.modals.login.loginModal.username', defaultMessage: 'User Name'})}
-                                        onChange={this.setUsername}
-                                        value={this.state.username}
-                                        onClick={(e: React.MouseEvent<HTMLInputElement>) => {
-                                            e.stopPropagation();
-                                        }}
-                                    />
-                                </Col>
-                            </FormGroup>
-                            <FormGroup>
-                                <Col
-                                    as={FormLabel}
-                                    sm={2}
-                                >
-                                    <FormattedMessage
-                                        id='renderer.modals.login.loginModal.password'
-                                        defaultMessage='Password'
-                                    />
-                                </Col>
-                                <Col sm={10}>
-                                    <FormControl
-                                        type='password'
-                                        placeholder={intl.formatMessage({id: 'renderer.modals.login.loginModal.password', defaultMessage: 'Password'})}
-                                        onChange={this.setPassword}
-                                        value={this.state.password}
-                                        onClick={(e: React.MouseEvent<HTMLInputElement>) => {
-                                            e.stopPropagation();
-                                        }}
-                                    />
-                                </Col>
-                            </FormGroup>
-                            <FormGroup>
-                                <Col sm={12}>
-                                    <div className='pull-right'>
-                                        <Button
-                                            type='submit'
-                                            variant='primary'
-                                        >
-                                            <FormattedMessage
-                                                id='label.login'
-                                                defaultMessage='Login'
-                                            />
-                                        </Button>
-                                        { ' ' }
-                                        <Button
-                                            variant='link'
-                                            onClick={this.handleCancel}
-                                        >
-                                            <FormattedMessage
-                                                id='label.cancel'
-                                                defaultMessage='Cancel'
-                                            />
-                                        </Button>
-                                    </div>
-                                </Col>
-                            </FormGroup>
-                        </Form>
-                    </Modal.Body>
-                </Modal>
-            </IntlProvider>
+            <Modal
+                show={Boolean(this.state.request && this.state.authInfo)}
+            >
+                <Modal.Header>
+                    <Modal.Title>
+                        <FormattedMessage
+                            id='renderer.modals.login.loginModal.title'
+                            defaultMessage='Authentication Required'
+                        />
+                    </Modal.Title>
+                </Modal.Header>
+                <Modal.Body>
+                    <p>
+                        {this.renderLoginModalMessage()}
+                    </p>
+                    <Form
+                        onSubmit={this.handleSubmit}
+                    >
+                        <FormGroup>
+                            <Col
+                                as={FormLabel}
+                                sm={2}
+                            >
+                                <FormattedMessage
+                                    id='renderer.modals.login.loginModal.username'
+                                    defaultMessage='User Name'
+                                />
+                            </Col>
+                            <Col sm={10}>
+                                <FormControl
+                                    type='text'
+                                    placeholder={intl.formatMessage({id: 'renderer.modals.login.loginModal.username', defaultMessage: 'User Name'})}
+                                    onChange={this.setUsername}
+                                    value={this.state.username}
+                                    onClick={(e: React.MouseEvent<HTMLInputElement>) => {
+                                        e.stopPropagation();
+                                    }}
+                                />
+                            </Col>
+                        </FormGroup>
+                        <FormGroup>
+                            <Col
+                                as={FormLabel}
+                                sm={2}
+                            >
+                                <FormattedMessage
+                                    id='renderer.modals.login.loginModal.password'
+                                    defaultMessage='Password'
+                                />
+                            </Col>
+                            <Col sm={10}>
+                                <FormControl
+                                    type='password'
+                                    placeholder={intl.formatMessage({id: 'renderer.modals.login.loginModal.password', defaultMessage: 'Password'})}
+                                    onChange={this.setPassword}
+                                    value={this.state.password}
+                                    onClick={(e: React.MouseEvent<HTMLInputElement>) => {
+                                        e.stopPropagation();
+                                    }}
+                                />
+                            </Col>
+                        </FormGroup>
+                        <FormGroup>
+                            <Col sm={12}>
+                                <div className='pull-right'>
+                                    <Button
+                                        type='submit'
+                                        variant='primary'
+                                    >
+                                        <FormattedMessage
+                                            id='label.login'
+                                            defaultMessage='Login'
+                                        />
+                                    </Button>
+                                    { ' ' }
+                                    <Button
+                                        variant='link'
+                                        onClick={this.handleCancel}
+                                    >
+                                        <FormattedMessage
+                                            id='label.cancel'
+                                            defaultMessage='Cancel'
+                                        />
+                                    </Button>
+                                </div>
+                            </Col>
+                        </FormGroup>
+                    </Form>
+                </Modal.Body>
+            </Modal>
         );
     }
 }
